CIL-2056 - Skip CoPerson records without Identifier
There are a few CoPerson records in https://registry-test.access-ci.org/
that do not have good data associated with them. In particular, the
provisioningData array does not have an "Identifier" index which causes
the syncPerson function to error out. This in turn causes the
AccessdbProvisioner to get stuck. So, check for problematic
provisioningData elements and simply log an error message.